+++ src/utils.jl
@@ -13,20 +13,6 @@ function emptybatch(model::T) where {T<:AbstractModel}
     end
 end
 
-"""
-    flipdict(dict::Dict)
-
-Flip the key=>value pair for each entry in a dict.
-"""
-flipdict(dict::Dict) = Dict(map(x->(dict[x],x),collect(keys(dict))))
-
-"""
-    createcountdict(dict::Dict)
-
-Create a dict copy of namesdict, for counting the occurances of each named object.
-"""
-createcountdict(dict::Dict) = Dict(map(x->(x,0),collect(keys(dict))))
-
 function gen_class_colors(model::YOLO.Yolo)
     classes = get_cfg(model)[:output][1][:classes]
     seed = [RGB{N0f8}(0,0,0), RGB{N0f8}(1,1,1)]
@@ -60,6 +46,25 @@ function _promote_to_n0f8(img)
     end
 end
 
+# Map normalized model-space bbox coordinates to image pixel scale, accounting
+# for the transpose between julia (column-major) and darknet (row-major) layouts
+function _box_geometry(img, model, transpose)
+    imgratio = size(img,2) / size(img,1)
+    if transpose
+        modelratio = get_cfg(model)[:width] / get_cfg(model)[:height]
+        idxs = (1, 2, 3, 4)
+    else
+        modelratio = get_cfg(model)[:height] / get_cfg(model)[:width]
+        idxs = (2, 1, 4, 3)
+    end
+    if modelratio > imgratio
+        h, w = size(img,1) .* (1, modelratio)
+    else
+        h, w = size(img,2) ./ (modelratio, 1)
+    end
+    return w, h, idxs
+end
+
 """
     draw_boxes(img::Array, model::YOLO.Yolo, padding::Array, results)
     draw_boxes!(img::Array, model::YOLO.Yolo, padding::Array, results)
@@ -81,19 +86,7 @@ function draw_boxes!(img::Union{Matrix{RGBA{N0f8}},Matrix{RGB{N0f8}}}, model::YO
         label_colors = gen_class_colors(model)
     end
 
-    imgratio = size(img,2) / size(img,1)
-    if transpose
-        modelratio = get_cfg(model)[:width]  / get_cfg(model)[:height]
-        x1i,y1i,x2i,y2i = 1,2,3,4
-    else
-        modelratio = get_cfg(model)[:height] / get_cfg(model)[:width]
-        x1i,y1i,x2i,y2i = 2,1,4,3
-    end
-    if modelratio > imgratio
-        h, w = size(img,1) .* (1, modelratio)
-    else
-        h, w = size(img,2) ./ (modelratio, 1)
-    end
+    w, h, (x1i, y1i, x2i, y2i) = _box_geometry(img, model, transpose)
     length(results) == 0 && return img
 
     img_rgb24 = similar(img, RGB24)
@@ -152,23 +145,7 @@ end
 
 # keep this for users that want to keep drawing boxes directly into non-color type images
 function draw_boxes!(img::AbstractArray, model::YOLO.Yolo, padding::AbstractArray, results; transpose=true, kwargs...)
-    imgratio = size(img,2) / size(img,1)
-    if transpose
-        modelratio = get_cfg(model)[:width] / get_cfg(model)[:height]
-        x1i, y1i, x2i, y2i = [1, 2, 3, 4]
-    else
-        modelratio = get_cfg(model)[:height] / get_cfg(model)[:width]
-        x1i, y1i, x2i, y2i = [2, 1, 4, 3]
-    end
-    if modelratio > imgratio
-        h, w = size(img,1) .* (1, modelratio)
-    else
-        h, w = size(img,2) ./ (modelratio, 1)
-    end
-
-    # p1 = Point(1, 1)
-    # p2 = Point(round(Int, w-((padding[x1i]+padding[x2i])*w)), round(Int, h-((padding[y1i]+padding[y2i])*h)))
-    # draw!(img, LineSegment(p1, p2), zero(eltype(img)))
+    w, h, (x1i, y1i, x2i, y2i) = _box_geometry(img, model, transpose)
     length(results) == 0 && return img
     for i in 1:size(results,2)
         bbox = results[1:4, i] .- padding
@@ -182,63 +159,4 @@ function draw_boxes!(img::AbstractArray, model::YOLO.Yolo, padding::AbstractArra
         draw!(img, pol, zero(eltype(img)))
     end
     return img
-end
-
-"""
-    benchmark(;select = [1,2,6], reverseAfter:Bool=false)
-
-Convenient benchmarking
-"""
-function benchmark(;select = [1,2,3,4,6,7,8,9], reverseAfter::Bool = false, img = rand(RGB,416,416), verbose=true, kw...)
-    pretrained_list = [
-                        YOLO.v2_tiny_416_COCO,
-                        YOLO.v3_tiny_416_COCO,
-                        YOLO.v4_tiny_416_COCO,
-                        YOLO.v7_tiny_416_COCO,
-                        YOLO.v2_416_COCO,
-                        YOLO.v3_416_COCO,
-                        YOLO.v3_spp_416_COCO,
-                        YOLO.v4_416_COCO,
-                        YOLO.v7_416_COCO,
-                        ][select]
-    reverseAfter && (pretrained_list = vcat(pretrained_list, reverse(pretrained_list)))
-
-
-    header = ["Model", "loaded?", "load time (s)", "#results", "run time (s)", "run time (fps)", "allocations"]
-    table = Array{Any}(undef, length(pretrained_list), 7)
-    for (i, pretrained) in pairs(pretrained_list)
-        modelname = string(pretrained)
-        verbose && @info "Loading and running $modelname"
-        table[i,:] = [modelname false "-" "-" "-" "-" "-"]
-
-        loaded = true
-        t_load = @elapsed begin
-            mod = try
-                pretrained(;silent=true, kw...)
-            catch ex
-                loaded = false
-                @warn "Failed to load $modelname: $ex"
-            end
-        end
-        table[i, 2] = loaded
-        loaded || continue
-
-        table[i, 3] = round(t_load, digits=3)
-
-        batch = emptybatch(mod)
-        batch[:,:,:,1], padding = prepare_image(img, mod)
-
-        res = mod(batch; detect_thresh=0.0, overlap_thresh=1.0) #run once
-        t_run = @belapsed $mod($batch; detect_thresh=0.0, overlap_thresh=1.0);
-        t_allocs = @allocated mod(batch; detect_thresh=0.0, overlap_thresh=1.0)
-        table[i, 4] = size(res, 2)
-        table[i, 5] = round(t_run, digits=4)
-        table[i, 6] = round(1/t_run, digits=1)
-        table[i, 7] = Base.format_bytes(t_allocs)
-
-        mod = nothing
-        batch = nothing
-        GC.gc()
-    end
-    pretty_table(table, header = header)
 end